Created new class (not yet implemented) AStarAlgorithm. Created new class PathFindingParameters. Updated PriorityQueue to use `int' priorities instead of 'float'.

+++ CoreOpenTibiaServer/src/COTS.GameServer/World/PathFinding/AStartNode.cs
@@ -3,7 +3,7 @@
     public sealed class AStartNode {
         public readonly uint X;
         public readonly uint Y;
-        public float Priority;
+        public int Priority;
         public int QueueIndex;
     }
 }
